Separating theme into theme file
