+++ package.json
@@ -10,9 +10,9 @@
     "core-js": "^2.5.7",
     "grommet": "^2.14.0",
     "html2pdf.js": "^0.9.1",
-    "lint-staged": "^7.0.4",
     "i18next": "^19.5.4",
     "i18next-browser-languagedetector": "^5.0.0",
+    "lint-staged": "^7.0.4",
     "node-sass-chokidar": "^1.2.2",
     "prettier": "1.12.1",
     "prop-types": "^15.7.2",
@@ -20,8 +20,9 @@
     "react": "^16.8.6",
     "react-async-component": "^2.0.0",
     "react-dom": "^16.8.6",
-    "react-is": "^16.13.1",
     "react-i18next": "^10.0.0",
+    "react-is": "^16.13.1",
+    "react-loader-spinner": "^3.1.14",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^3.3.0",
     "react-stepper-horizontal": "^1.0.10",
