+++ src/components/async-download-pdf/index.js
@@ -3,6 +3,8 @@ import * as React from 'react'
 import {asyncComponent} from 'react-async-component'
 import DownloadPDF from './DownloadPDF'
 import {withTranslation} from 'react-i18next'
+import Loader from 'react-loader-spinner'
+import { POA_CORAL_BLUE } from '../../styles/grommet/theme'
 
 // TODO Toast for error
 const ErrorNotification = ({t}) => (
@@ -12,7 +14,12 @@ const ErrorNotification = ({t}) => (
 // TODO Add spinner back for loading component
 const asyncDownload: typeof DownloadPDF = asyncComponent({
   ErrorComponent: withTranslation()(ErrorNotification),
-  LoadingComponent: () => (<div>Please wait...</div>),
+  LoadingComponent: () =>
+    <Loader
+      type="Oval"
+      color={POA_CORAL_BLUE}
+      height={80}
+      width={80}/>,
   name: 'AsyncDownloadPDF',
   resolve: () => import('./DownloadPDF')
 })
